Refactoring of a few BigQuery classes.
Also adding RowAsDictJsonCoder in bigquery.py for backwards compatibility
